Remove dt::array class (#2341)
In this PR we delete the `dt::array` class, replacing it with `Buffer` everywhere (in one place replaced with `std::vector`).
This class was mainly used in `RowIndex`, however it was not possible to cleanly convert such array into a `Buffer` and back. By using `Buffer` s everywhere we should now avoid some of the unnecessary data copies.