Frame.source (#2393)
Added new frame property: Frame.source. This contains the name of the file where the frame was loaded from. In some other cases it may contain a generic descriptor: "<file>", "<text>", "<pickle>" if the source cannot be identified more precisely.

The property is read-only. It will clear if the frame is modified after being loaded.