Fix type bump message generation (#2401)
`fread` could generate type bump messages that contain a sample of data. To limit the size of the message we have used `BUF_SIZE` variable, but this approach doesn't work well when the data size becomes sufficiently large. 

In this PR we fix the type bump message generation by applying the `BUF_SIZE` (now it is called `MAX_FIELD_LEN`) to the data only and not to the whole message. In such a way the other useful information in the message never gets truncated.

Also fix several issues introduced in the previous PR's and several larger fread tests.

WIP for #2301