Use platform specific tolerance in certain fread tests (#2408)
Not all the platforms have a proper long double type that is used by fread to convert double literals into double numbers. In this PR we make use of platform specific tolerances for Windows and PowerPC64 that allow certain fread tests to not fail. 

Also, set up AppVeyor config file to launch tests for every Windows build.

WIP for #2301