Disable freading from file-like objects on Windows (#2412)
Right now to read data from a file object, datatable needs a file descriptor to be used on the C++ side. Since file descriptors are not native to all the platforms, this functionality doesn't work on Windows. Some workarounds are possible, however, since this feature is not so widely used, in this PR we temporarily disable it on Windows and throw `NotImplementedError` instead.

Closes #2301