Fread PreColumn refactoring (#2445)
This is an attempt to split the PreColumn class into 2 components: an InputColumn and an OutputColumn.

In a typical scenario, there is a 1-to-1 correspondence between these. When the user requests that only some columns were read, then the set of OutputColumns becomes smaller, currently we handle this via a simple flag present_in_output_. However, there are more advanced cases: output columns that require pre-processing of the inputs.

This PR does not attain full independence between InputColumn and OutputColumn. However, it has the advantage of not changing any of the existing functionality.