Wheel generation adjustments (#2442)
As of this PR, instead of the Jenkins internal counter
- master builds will use `git rev-list --count master` as a `DT_BUILD_NUMBER`
- other branches will use `git rev-list --count master..` in the `DT_BUILD_SUFFIX`

This is now valid for both AppVeyor and Jenkins making the wheel names consistent. Travis still uses its own naming scheme. 

In addition, we now do not use `scm` when checking out the code on Jenkins and do a simple `git clone` instead. This makes the structure of the source code directory cleaner and consistent with github.

Closes https://github.com/h2oai/datatable/issues/2440