Fix slow reading of string columns by fread (#2461)
The issue was that the mutex in `MemoryWritableBuffer` was acquired too often, causing major slowdown as multiple threads were trying to access it at the same time. The solution was to create a helper `Writer` class that acquires the mutex for the duration of multiple writes (note: this doesn't cause unnecessary contention, since the lock is acquired in "read-only" mode, and multiple threads can do that at the same time).

In addition, `ParseContext` now uses simple `Buffer` instead of `MemoryWritableBuffer` for its string content. That buffer is cleared when we start reading a chunk.

Also: `ParseContext.anchor` was removed, since it is no longer needed.

Closes #2458