Exception initialization is now done at datatable import (#2527)
Previously, we had function `init()` in exceptions.cc trying to initialize exception objects when they were first needed (i.e. when an exception was first thrown). However, because this was not done in thread-safe manner, it was possible to get a situation when multiple threads are trying to initialize those exception objects at the same time, resulting in undefined behavior, up to and including a segfault.

In order to avoid this, now we initialize all exception objects upon module import (in the special "after-import" stage), when no threads are in play, and consequently there is no possibility for any race conditions.

In addition, initialization of python `options` object moved to `__init__.py` in order to make the initialization sequence more well-defined.

Closes #2526