Script for checking standard library headers (#2529)
Created a script that can scan all source files for the presence of std:: symbols, and then check whether the corresponding standard library headers were included.

This can be used to detect the situation where a certain symbol from the standard library is used without the header, and it isn't detected by the compiler because that symbol is transitively included from some other standard library. However, when trying to compile with a different compiler that uses different standard library version or implementation that symbol may end up no longer transitively included, resulting in a compilation error.

Previously we hit this problem multiple times, and so this script is intended to prevent it from happening.

The script needs to be run manually. It contains the list of "known" standard library symbols and their corresponding header files, which may need to be updated in the future if we use any symbol that is not in the list.