Created ..xcode::shell Sphinx directive (#2531)
The new directive should be used for all code blocks with shell commands, e.g:
```rst
.. xcode:: shell

    $ pip install datatable
```
Unlike the traditional `.. code-block::` we have full control over lexing / styling the final output. In particular, with this patch the shell prompt is distinct from the rest of the command line; and it is also marked with `user-select: none` in CSS, meaning it won't get in the way when a user tries to copy the commands.

Most importantly, we will extend the `.. xcode::` directive in the future to python languages, in particular allowing for automatic detection of datatable frames in the output.