Fix loading a Frame from a temporary bytes Jay object (#2548)
Before, when fread was given a string/bytes object, we were creating a `Buffer::external` view on that object, and then read that data. This worked fine when reading CSVs, but if the file in fact is a Jay file, then we end up creating columns that are direct views on that data. These columns are then put together in a Frame and returned to the user. The problem is that `Buffer::external` is not designed for long-term data storage: it does not own its memory region. So, when the original bytes object is gc'd, we end up with a Frame that contains dangling pointers. 

Now, I've created a new type of `Buffer`: `PyBytes_BufferImpl`, which keeps the source py object inside it, ensuring that the memory region wouldn't get deallocated prematurely.

The impact of this bug is uncler: on one hand, users usually save Jay into a file instead of a bytes object; on the other hand, pickling might in theory do exactly that.

Closes #2547