New framework for writing C++ tests (#2557)
Created a new framework for writing C++ tests:
  - a test can be declared with a simple `TEST(suite, name)` macro;
  - no need to write extra code to make the test available in python: the test will get discovered and run in pytest automatically (provided it is added to an existing suite, in order to handle a new suite a 3-line pytest function must be added);
  - added a set of macros for writing assertions; when failing these macros will report the line number and values of their arguments:
    - `ASSERT_EQ`, `ASSERT_NE`, `ASSERT_LT`, `ASSERT_LE`, `ASSERT_GT`, `ASSERT_GE`;
    - `ASSERT_FLOAT_EQ` (approximate comparison of doubles);
    - `ASSERT_TRUE`, `ASSERT_FALSE`;
    - `ASSERT_THROWS` (with optional checks for exception class and message);
  - class `Error` merged with `PyError` and `Warning` in order to simplify dealing with exceptions;
  - existing C++ tests were cleaned up and converted into the new framework.

Generally, this framework is not too far removed from GoogleTest; main difference is that we aim to integrate with pytest directly, whereas GoogleTest would require all tests to be run from a separate executable. Our approach is more flexible, as we can select which subset of tests to run using standard pytest's syntax, and also the tests can be easily combined with python code. 

In the future we may also add parametrized tests, similar to pytest's approach.