Do not test API docs if API path doesn't exist (#2648)
When we do `make sdist` we do not include the `docs/api` folder, but we do include all the tests meaning that the `test-docs.py` is also included. Since we don't want this test to fail for no reason, in this PR we disable it, as it will otherwise fail when no `docs/api` folder is present on the system.