Automatically build documentation on Travis (#2745)
This PR creates a test that would  build the documentation with Sphinx and  check that there are no warnings/errors. Since  this test is time-consuming, it needs to be enabled via an environment variable `DT_BUILD_DOCS`. At the moment I've enabled this test on Travis for python 3.6 (it appears that sphinx is much slower under python 3.8).

Closes #2744