Create new column type: VOID (#2783)
The type "void" describes a column that has only NA values. For example, a python list `[None, None, None, ...]` creates a  column of type void. This type corresponds to arrow `null` type, or to numpy's `void`.

This type is upward-compatible with any other type, in the sense that it can be automatically promoted to any other type in any operation that demands it.

Before this PR, the stype `VOID` already existed internally. However, it was used as a special "fake" stype to indicate various conditions. With this PR we introduce two new special stypes: `INVALID` and `AUTO`, which are used internally only.

Closes #2666