New Type class (#2849)
The class Type is intended as a replacement for stype/ltype enums.
The main difference from `stype` is that the new class is not an enum, and supports arbitrary type descriptors.

This new class is necessary in order to implement datetime types with various precision / timezone parameters. In the future this class will also be able to accommodate columns of "list" types, compound (struct) columns, union columns, etc.

In the future we want to gradually replace all uses of `stype` with the equivalent code that uses `Type`s.