Refactor row* functions to use FExprs (#2872)
Each row* function now supports vararg-style invocation, e.g. `rowsum(f.A, f.B, f.C)`, in addition to the regular single-argument style. On the other hand, applying row-function to a frame no longer produces a frame -- instead, it's an FExpr. This functionality previously existed but was undocumented. It would be better to return this functionality as a method on `Frame` class, instead of being a special case for the row* functions.

WIP for #2562