Do data shuffling when fitting a linear model (#2914)
In general, it is recommended to shuffle training data when fitting the model with SGD. The same is suggested in the [paper](http://martin.zinkevich.org/publications/nips2010.pdf) that we base our parallel implementation on. 

In this PR we implement quasi-random data shuffling for the training data, similar to what we currently do in Aggregator. The shuffling could be disabled by setting the seed to zero.

WIP for #2358