py::XArgs can now be used to declare methods (#2948)
At some point, `XArgs` should be able to completely replace `PKArgs`.
This PR makes it possible for `XArgs` to be used in methods, with the help of macros `DECLARE_METHOD` and `DECLARE_METHODv` (the latter is for methods returning `void`). These macros provide automatic registration of methods with their python class, avoiding the need to declare the methods in multiple places.

Some methods of `py::Frame` were converted into the new format.