fread parse error (#2708)
At `src/core/read/fread/fread_thread_context.cc:210,` If `j<ncols` and if `*tch` reaches the end of line with the current quote rule, the while loop exits ending up in IOError. 

With the code added in this PR, one way to fix this is to decrement tch`(tch--)` to the previous `sep` and try parsing again with the next quote rule. If the number of columns in the current row is only one, decrementing tch will never hit the sep in the current row and end up running into the previous row, hence the `if` logic inside the `while` loop to prevent this.

Closes #2680