Visually separate Type variables from properties/methods (#2993)
So, the problem with documenting `Type` class is that it has static members such as `Type.str32` or `Type.float64`, and also regular properties such as `Type().name` or `Type().max`. These obviously have different usage (although technically static members are accessible on instances too, it's not how they are supposed to be used), and therefore mixing them up in the documentation creates lots of confusion. 

This PR addresses this problem by adding a small star icon next to each static member, to indicate that they are "special".