String slicing (#3015)
This implements string-slicing function `FExpr[a:b:c]`, or alternatively as a regular function `dt.str.slice(col, start=a, stop=b, step=c)`.

The new function supports both constant slices and expression-slices. 

The reason to have both a function and a slice expression `[a:b]` is that a function is much easier to find in the documentation (esp. once we have more string functions in `dt.str` module), whereas `[]`-based slices are easier to write.

Closes #1667