Fix latent column materialization routine (#3102)
Latent column materialization routine was missing logic for setting up the `refcount` property. In this PR we fix it, so that the newly created material column gets the same `refcount` as had the original latent column. In some cases, like for the recently added `qcut` functionality, having wrong `refcount` could result in segfaults.

Closes #3061