Fix `.stype.dtype` when called on categorical and array types (#3206)
Even though we're going to deprecate `.stype` property, we still don't want `.stype.dtype` to produce unhandled exceptions for the new types. In this PR we add categorical and array types to  `_stype_2_dtype` to prevent errors.

We also add a test for `to_numpy()` when called on datetime columns.