Switch to `manylinux2014` image on `x86_64` and to py38 for debug wheels (#3327)
Support for `manylinux2010` image that we're using to build datatable on `x86_64` is about to be dropped (https://github.com/pypa/manylinux/issues/1281), so we switch to `manylinux2014` that we're already using on `ppc64le`. 

Also, Python 3.7 will reach its end of life soon, hence we switch to Python 3.8 when generating debug wheels. In principle, we can generate debug wheels for all the supported Python versions, however, this will significantly slow down our building pipeline.