Fix cumulative functions when invoked on the latent columns (#3348)
Cumulative functions acces the column's data in parallel. If the input column appears to be a `Latent_ColumnImpl`, materialization of such a column happens in parallel. However, the `Latent_ColumnImpl::materialize()` is not a thread-safe method and invoking it from multiple threads results in a segfault.

In this PR we `vivify()` the source column, i.e. materialization, if necessary, happens in one thread only, before processing it in parallel.

Closes #3345