Fix chained `FExpr` reducers (#3446)
Chained reducers is a pretty rare use case, because the second reducer gets a scalar (per group) as an input. Some SQL engines don't even allow that. However, we never seems to forbid chaining explicitly, at the same time, we had no tests and didn't think such a use case through. As a result, we either produced wrong results (see #3417) or even segfaulted when chaining was involved. In this PR we fix all the known related issues and add some tests. We also

- refactor `FExpr` reducers to inherit from `FExpr_ReduceUnary`;
- improve reducers logic with respect to grouped/non-grouped frames.

WIP for #3417